refactor kilt dataset
