+++ config/dataset/kilt_aidayago2.yaml
@@ -9,7 +9,7 @@ dev:
     
     query: 
         init_args:
-            _target_: modules.dataset_processor.KILTAidayago2
+            _target_: modules.processors.kilt_dataset_processor.KILTAidayago2
             split: "validation"
 test:
     doc: null
