+++ config/dataset/kilt_cweb.yaml
@@ -9,7 +9,7 @@ dev:
     
     query: 
         init_args:
-            _target_: modules.dataset_processor.KILTCweb
+            _target_: modules.processors.kilt_dataset_processor.KILTCweb
             split: "validation"
 test:
     doc: null
