+++ config/dataset/kilt_eli5.yaml
@@ -9,7 +9,7 @@ dev:
     
     query: 
         init_args:
-            _target_: modules.dataset_processor.KILTEli5
+            _target_: modules.processors.kilt_dataset_processor.KILTEli5
             split: "validation"
 test:
     doc: null
