+++ config/dataset/kilt_fever.yaml
@@ -9,7 +9,7 @@ dev:
     
     query: 
         init_args:
-            _target_: modules.dataset_processor.KILTFever
+            _target_: modules.processors.kilt_dataset_processor.KILTFever
             split: "validation"
 test:
     doc: null
