+++ config/dataset/kilt_structured_zeroshot.yaml
@@ -9,7 +9,7 @@ dev:
     
     query: 
         init_args:
-            _target_: modules.dataset_processor.KILTStructuredZeroshot
+            _target_: modules.processors.kilt_dataset_processor.KILTStructuredZeroshot
             split: "validation"
 test:
     doc: null
