+++ modules/dataset_processor.py
@@ -138,260 +138,6 @@ class BIOASQ11B(Processor):
         return dataset
     
     
-class TimeSensitiveQA(Processor):
-
-    def __init__(self,  *args, **kwargs):
-        self.dataset_name = 'TimeSensitiveQA'
-        #self.path = data_path
-        super().__init__(*args, **kwargs, dataset_name=self.dataset_name)
-    
-    def process(self):
-        hf_name = "diwank/time-sensitive-qa"
-        #dataset = datasets.load_dataset("json", data_files=[self.path])[self.split]
-        dataset = datasets.load_dataset(hf_name, num_proc=self.num_proc)[self.split]
-        #['id','docs','question','type','ideal_answer','exact_answer','snippets']
-        dataset = dataset.map(lambda example: {'label': example['targets']})
-        dataset = dataset.rename_column("question", "content")
-        dataset = dataset.rename_column("idx", "id")
-        dataset = dataset.remove_columns(['context', 'paragraphs'])
-        return dataset
-    
-
-class WIKIQA(Processor):
-    def __init__(self, *args, **kwargs):
-        dataset_name = 'wiki_qa'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    def process(self):
-        hf_name = 'wiki_qa' 
-        dataset = datasets.load_dataset(hf_name, num_proc=self.num_proc)[self.split]
-        # discarding empty answers 
-        dataset_f = dataset.filter (lambda x: x['label'] == 1) # keeping only the valid sentences
-        
-        # ranking_label: list of wikipedia_ids per answer, empty list if no provenances are present or answer is empty
-        # No ranking labels
-        #dataset = dataset.map(lambda example: {'ranking_label': [[provenance['wikipedia_id'] for provenance in el['provenance']] if len(el['answer']) > 0 and len(el['provenance']) > 0 else [] for el in example['output']]})
-
-        dataset_l=[]
-
-        qid_set=set(dataset_f['question_id'])
-
-        for q in qid_set:
-            qsel= dataset_f.filter(lambda x: x['question_id']==q)
-            ex={ 'id':q ,
-                 'content':qsel['question'][0],
-                  'label': qsel['answer']}
-            dataset_l.append(ex)
-        
-        #dataset = dataset_f.rename_column("question", "content")
-        #dataset = dataset.remove_columns(['document_title', 'label'])
-        #dataset = dataset.rename_column("answer", "label")
-        #dataset = dataset.rename_column("question_id", "id")
-        dataset=datasets.Dataset.from_list(dataset_l)
-        return dataset
-
-class SCIQ(Processor):
-    def __init__(self, *args, **kwargs):
-        dataset_name = 'sciq'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    def process(self):
-        hf_name = 'sciq' 
-        dataset = datasets.load_dataset(hf_name, num_proc=self.num_proc)[self.split]
-        
-        ###dataset = dataset.map(lambda example: {'label': [example['correct_answer']]})
-        #dataset = dataset.remove_columns(["correct_answer","distractor1", "distractor2","distractor3","support"])
-        #generating the id, train_0 ... validation_0 validation_1
-        cid= [self.split+str(i) for i in range(len(dataset))]
-        dataset = dataset.add_column("id", cid)
-        if self.oracle_provenance:
-            # document
-            dataset = dataset.rename_column('support','content')
-            dataset = dataset.remove_columns(["question","correct_answer","distractor1", "distractor2","distractor3"])            
-            #dataset = datasets.Dataset.from_dict({'content': paragraphs, 'id': ids})
-        else:
-            # query
-            dataset = dataset.rename_column("question", "content")
-            dataset = dataset.map(lambda example: {'label': [example['correct_answer']]})
-            dataset = dataset.remove_columns(["support","correct_answer","distractor1", "distractor2","distractor3"])        
-        return dataset
-
-
-
-
-class ASQA(Processor):
-    wiki_api = "https://en.wikipedia.org/w/api.php?action=query&format=json&titles={}"
-    def __init__(self, *args, **kwargs):
-        dataset_name = 'asqa'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    @staticmethod
-    def fetch_wiki_id(inp):
-        wiki_link, wiki_title = inp
-        response = requests.get(ASQA.wiki_api.format(wiki_title))
-        wiki_object_has_no_pages = 0
-        if response.status_code == 200:
-            data = response.json()
-            # hack because of weird dict structure returned json of the wiki_api
-            pages = data.get("query", {}).get("pages", {})
-            if pages:
-                wiki_id = next(iter(pages.keys()))
-                #return (wiki_link, wiki_id)
-                return wiki_id
-            else:
-                wiki_object_has_no_pages += 1
-                return None
-        else:
-            count_not_found += 1
-            print(f"wiki page {wiki_link} could not be fetched from wiki!")
-            return None
-
-
-    def process(self):
-
-        hf_name = 'din0s/asqa' 
-        dataset = datasets.load_dataset(hf_name, num_proc=self.num_proc)[self.split]
-        
-        #features: ['ambiguous_question', 'qa_pairs', 'wikipages', 'annotations', 'sample_id'],
-
-        # use "sample_id" ?
-        dataset = dataset.map(lambda example, idx: {'id': str(idx), **example}, with_indices=True)
-
-        #dataset = dataset.rename_column("answer", "label")
-        dataset = dataset.rename_column("ambiguous_question", "content")
-
-        # get short answers
-        def short_answers(example):
-            #z=[ q for ex in dataset['dev'][:2]['qa_pairs']  for x in ex for q in x['short_answers']]~
-            z= list(set([ ans for qa in example['qa_pairs']  for ans in qa['short_answers'] ]))
-            # or z=[ x['short_answers']  for ex in qa_pairs for x in ex]
-            return z #[pair[] for pair in qa_pairs]
-        def get_wiki_id(example):
-                wiki_ids=list()
-                wiki_objects = example['wikipages']
-                for wiki_object in wiki_objects:
-                        if wiki_object['url'] != None:
-                            # get wiki url
-                            wiki_link = wiki_object['url']
-                            # get title
-                            wiki_title = wiki_link.split("/")[-1]
-                            # fetch id by title
-                            wiki_ids.append(ASQA.fetch_wiki_id((wiki_link, wiki_title)))
-                return wiki_ids
-
-
-        # long_awser ?
-        #  example['annotation'][0]['long_answer']
-        # Apply the cleaning function to the 'label' column
-        dataset = dataset.map(lambda example: {'label': short_answers(example)})
-        #dataset = dataset.map(lambda example: {'ranking_label': get_wiki_id(example)},num_proc=5)
-
-        dataset = dataset.remove_columns([ 'qa_pairs', 'wikipages', 'annotations', 'sample_id'])
-
-        # ranking_label: wikipedia url
-
-        return dataset
-        
-
-#truthful_qa
-class truthful_qa(Processor):
-    """
-    DatasetDict({
-    validation: Dataset({
-        features: ['type', 'category', 'question', 'best_answer', 'correct_answers', 'incorrect_answers', 'source'],
-        num_rows: 817
-    })
-})
-
-    """
-    wiki_api = "https://en.wikipedia.org/w/api.php?action=query&format=json&titles={}"
-    def __init__(self, *args, **kwargs):
-        dataset_name = 'truthful_qa'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    @staticmethod
-    def fetch_wiki_id(inp):
-        wiki_link, wiki_title = inp
-        response = requests.get(truthful_qa.wiki_api.format(wiki_title))
-        wiki_object_has_no_pages = 0
-        if response.status_code == 200:
-            data = response.json()
-            # hack because of weird dict structure returned json of the wiki_api
-            pages = data.get("query", {}).get("pages", {})
-            if pages:
-                wiki_id = next(iter(pages.keys()))
-                return wiki_id
-            else:
-                wiki_object_has_no_pages += 1
-                return ""
-        else:
-            count_not_found += 1
-            print(f"wiki page {wiki_link} could not be fetched from wiki!")
-            return ""
-
-
-    def process(self):
-
-        hf_name = 'truthful_qa' 
-        dataset = datasets.load_dataset(hf_name, "generation",num_proc=self.num_proc)[self.split]
-        
-
-        # use "sample_id" ?
-        dataset = dataset.map(lambda example, idx: {'id': str(idx), **example}, with_indices=True)
-
-        # dataset = dataset.rename_column("best_answer", "label")
-        dataset = dataset.map(lambda example: {'label': [example['best_answer']]})
-        dataset = dataset.rename_column("question", "content")
-
-        def get_wiki_id(example):
-            wiki_link = example['source']
-            wiki_title = wiki_link.split("/")[-1]
-            id= truthful_qa.fetch_wiki_id((wiki_link, wiki_title))
-            return truthful_qa.fetch_wiki_id((wiki_link, wiki_title))
-
-        #dataset = dataset.map(lambda example: {'ranking_label': get_wiki_id(example)},num_proc=5)
-
-        dataset = dataset.remove_columns([ 'best_answer', 'type','category', 'correct_answers','incorrect_answers','source'])
-
-        # ranking_label: wikipedia url
-
-        return dataset
-        
-class POPQA(Processor):
-    def __init__(self, *args, **kwargs):
-        dataset_name = 'popqa'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    def process(self):
-        hf_name = 'akariasai/PopQA' 
-        dataset = datasets.load_dataset(hf_name, num_proc=self.num_proc)[self.split]
-        dataset = dataset.rename_column("question", "content")
-        dataset = dataset.map(lambda example: {'label': eval(example['possible_answers'])})
-        dataset = dataset.remove_columns(["possible_answers","id","subj", "prop","obj","subj_id","prop_id",'obj_id','s_aliases','o_aliases','s_uri','o_uri','s_wiki_title','o_wiki_title','s_pop','o_pop'])
-        #generating the id, train_0 ... validation_0 validation_1
-        cid= [self.split+str(i) for i in range(len(dataset))]
-        dataset = dataset.add_column("id", cid)
-
-        return dataset
-
-class wikimultihopqa(Processor):
-    def __init__(self, *args, **kwargs):
-        dataset_name = '2wikimultihopqa'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    def process(self):
-        hf_name = 'scholarly-shadows-syndicate/2wikimultihopqa_with_q_gpt35' 
-        dataset = datasets.load_dataset(hf_name, num_proc=self.num_proc)[self.split]
-        dataset = dataset.rename_column("question", "content")
-        #dataset = dataset.rename_column("answer", "label")
-        dataset = dataset.map(lambda example: {'label': [example['answer']]})
-        dataset = dataset.remove_columns(["answer","evidences", "supporting_facts","context"])
-        #generating the id, train_0 ... validation_0 validation_1
-        cid= [self.split+str(i) for i in range(len(dataset))]
-        dataset = dataset.add_column("id", cid)
-        
-        return dataset
-   
 
 class MMLU(Processor):
 
@@ -457,248 +203,6 @@ class NQOpen(Processor):
 
         return dataset
 
-class KILTNQ(Processor):
-
-    def __init__(self, *args, **kwargs):
-        dataset_name = 'kilt_nq'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    def process(self):
-        hf_name = 'kilt_tasks' 
-        hf_query_or_doc_name= "nq"
-        dataset = datasets.load_dataset(hf_name, hf_query_or_doc_name, num_proc=self.num_proc)[self.split]
-        # discarding empty answers 
-        dataset = dataset.map(lambda example: {'label': [el['answer'] for el in example['output'] if len(el['answer']) > 0]})
-
-        # ranking_label: list of wikipedia_ids per answer, empty list if no provenances are present or answer is empty
-        dataset = dataset.map(lambda example: {'ranking_label': [[provenance['wikipedia_id'] for provenance in el['provenance']] if len(el['answer']) > 0 and len(el['provenance']) > 0 else [] for el in example['output']]})
-        dataset = dataset.rename_column("input", "content")
-        dataset = dataset.remove_columns(['meta', 'output'])
-        return dataset
-
-
-class KILTTriviaqa(Processor):
-
-    def __init__(self, *args, **kwargs):
-        dataset_name = 'kilt_triviaqa'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    def process(self):
-        
-        # Get the KILT task datasets
-        dataset = datasets.load_dataset("kilt_tasks", name="triviaqa_support_only")[self.split]
-        hf_q_ids = set(dataset['id'])
-        # Most tasks in KILT already have all required data, but KILT-TriviaQA
-        # only provides the question IDs, not the questions themselves.
-        # Thankfully, we can get the original TriviaQA data with:
-        trivia_qa = datasets.load_dataset('trivia_qa', 'unfiltered.nocontext')[self.split]
-        # The KILT IDs can then be mapped to the TriviaQA questions with:
-
-        def add_missing_data(x, trivia_qa_subset, triviaqa_map):
-            i = triviaqa_map[x['id']]
-            x['input'] = trivia_qa_subset[i]['question']
-            x['output'][0]['original_answer'] = trivia_qa_subset[i]['answer']['value']
-            return x
-            
-        triviaqa_map = dict([(q_id, i) for i, q_id in enumerate(trivia_qa['question_id'])])
-        dataset = dataset.filter(lambda x: x['id'] in triviaqa_map, num_proc=self.num_proc)
-        # only use ids that are present in the kilt_dataset
-        dataset = dataset.filter(lambda x: x['id'] in hf_q_ids, num_proc=self.num_proc)
-        dataset = dataset.map(add_missing_data, fn_kwargs=dict(trivia_qa_subset=trivia_qa, triviaqa_map=triviaqa_map), num_proc=self.num_proc)
-
-        # discarding empty answers 
-        dataset = dataset.map(lambda example: {'label': [el['answer'] for el in example['output'] if len(el['answer']) > 0]})
-        # ranking_label: list of wikipedia_ids per answer, empty list if no provenances are present or answer is empty
-        dataset = dataset.map(lambda example: {'ranking_label': [[provenance['wikipedia_id'] for provenance in el['provenance']] if len(el['answer']) > 0 and len(el['provenance']) > 0 else [] for el in example['output']]})
-        dataset = dataset.rename_column("input", "content")
-        dataset = dataset.remove_columns(['meta', 'output'])
-        return dataset
-
-class KILTHotpotqa(Processor):
-
-    def __init__(self, *args, **kwargs):
-        dataset_name = 'kilt_hotpotqa'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    def process(self):
-        hf_name = 'kilt_tasks' 
-        hf_query_or_doc_name= "hotpotqa"
-        dataset = datasets.load_dataset(hf_name, hf_query_or_doc_name, num_proc=self.num_proc)[self.split]
-        # discarding empty answers 
-        dataset = dataset.map(lambda example: {'label': [el['answer'] for el in example['output'] if len(el['answer']) > 0]})
-        # ranking_label: list of wikipedia_ids per answer, empty list if no provenances are present or answer is empty
-        dataset = dataset.map(lambda example: {'ranking_label': [[provenance['wikipedia_id'] for provenance in el['provenance']] if len(el['answer']) > 0 and len(el['provenance']) > 0 else [] for el in example['output']]})
-        dataset = dataset.rename_column("input", "content")
-        dataset = dataset.remove_columns(['meta', 'output'])
-        return dataset
-    
-class KILTAidayago2(Processor):
-
-    def __init__(self, *args, **kwargs):
-        dataset_name = 'kilt_aidayago2'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    def process(self):
-        hf_name = 'kilt_tasks' 
-        hf_query_or_doc_name= "aidayago2"
-        dataset = datasets.load_dataset(hf_name, hf_query_or_doc_name, num_proc=self.num_proc)[self.split]
-        # discarding empty answers 
-        dataset = dataset.map(lambda example: {'label': [el['answer'] for el in example['output'] if len(el['answer']) > 0]})
-        # ranking_label: list of wikipedia_ids per answer, empty list if no provenances are present or answer is empty
-        dataset = dataset.map(lambda example: {'ranking_label': [[provenance['wikipedia_id'] for provenance in el['provenance']] if len(el['answer']) > 0 and len(el['provenance']) > 0 else [] for el in example['output']]})
-        dataset = dataset.rename_column("input", "content")
-        dataset = dataset.remove_columns(['meta', 'output'])
-        return dataset
-    
-class KILTCweb(Processor):
-
-    def __init__(self, *args, **kwargs):
-        dataset_name = 'kilt_cweb'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    def process(self):
-        hf_name = 'kilt_tasks' 
-        hf_query_or_doc_name= "cweb"
-        dataset = datasets.load_dataset(hf_name, hf_query_or_doc_name, num_proc=self.num_proc)[self.split]
-        # discarding empty answers 
-        dataset = dataset.map(lambda example: {'label': [el['answer'] for el in example['output'] if len(el['answer']) > 0]})
-        # ranking_label: list of wikipedia_ids per answer, empty list if no provenances are present or answer is empty
-        dataset = dataset.map(lambda example: {'ranking_label': [[provenance['wikipedia_id'] for provenance in el['provenance']] if len(el['answer']) > 0 and len(el['provenance']) > 0 else [] for el in example['output']]})
-        dataset = dataset.rename_column("input", "content")
-        dataset = dataset.remove_columns(['meta', 'output'])
-        return dataset
-    
-class KILTEli5(Processor):
-
-    def __init__(self, *args, **kwargs):
-        dataset_name = 'kilt_eli5'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-
-    def process(self):
-        hf_name = 'kilt_tasks' 
-        hf_query_or_doc_name= "eli5"
-
-        dataset = datasets.load_dataset(hf_name, hf_query_or_doc_name, num_proc=self.num_proc)[self.split]
-
-        # hf dataset is missing provenances add them
-        if self.split == 'dev':
-            original_fname = '../data/eli5-dev-kilt.jsonl'
-            if not os.path.exists(original_fname):
-                urllib.request.urlretrieve("https://dl.fbaipublicfiles.com/KILT/eli5-dev-kilt.jsonl", original_fname)
-
-            original = defaultdict(dict)
-            for l in open(original_fname):
-                example = json.loads(l)
-                original_provenance = [out['provenance'] for out in example['output'] if 'provenance' in out]
-                q_id = example['id']
-                original[q_id] = original_provenance
-
-            def add_missing_data(x, original_dataset):
-                id_ = x['id']
-                original_provenance = original_dataset[id_]
-                for i in range(len(original_provenance)):
-                    x['output'][i]['provenance'] = original_provenance[i]
-                return x
-            
-            dataset = dataset.map(add_missing_data, fn_kwargs=dict(original_dataset=original))
-        dataset = dataset.map(lambda example: {'label': [el['answer'] for el in example['output'] if len(el['answer']) > 0]})
-        # ranking_label: list of wikipedia_ids per answer, empty list if no provenances are present or answer is empty
-        dataset = dataset.map(lambda example: {'ranking_label': [[provenance['wikipedia_id'] for provenance in el['provenance']] if len(el['answer']) > 0 and len(el['provenance']) > 0 else [] for el in example['output']]})
-        dataset = dataset.rename_column("input", "content")
-        dataset = dataset.remove_columns(['meta', 'output'])
-        return dataset
-    
-class KILTFever(Processor):
-
-    def __init__(self, *args, **kwargs):
-        dataset_name = 'kilt_fever'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    def process(self):
-        hf_name = 'kilt_tasks' 
-        hf_query_or_doc_name= "fever"
-        dataset = datasets.load_dataset(hf_name, hf_query_or_doc_name, num_proc=self.num_proc)[self.split]
-        # discarding empty answers 
-        dataset = dataset.map(lambda example: {'label': [el['answer'] for el in example['output'] if len(el['answer']) > 0]})
-        # ranking_label: list of wikipedia_ids per answer, empty list if no provenances are present or answer is empty
-        dataset = dataset.map(lambda example: {'ranking_label': [[provenance['wikipedia_id'] for provenance in el['provenance']] if len(el['answer']) > 0 and len(el['provenance']) > 0 else [] for el in example['output']]})
-        dataset = dataset.rename_column("input", "content")
-        dataset = dataset.remove_columns(['meta', 'output'])
-        return dataset
-    
-class KILTStructuredZeroshot(Processor):
-
-    def __init__(self, *args, **kwargs):
-        dataset_name = 'kilt_structured_zeroshot'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    def process(self):
-        hf_name = 'kilt_tasks' 
-        hf_query_or_doc_name= "structured_zeroshot"
-        dataset = datasets.load_dataset(hf_name, hf_query_or_doc_name, num_proc=self.num_proc)[self.split]
-        # discarding empty answers 
-        dataset = dataset.map(lambda example: {'label': [el['answer'] for el in example['output'] if len(el['answer']) > 0]})
-        # ranking_label: list of wikipedia_ids per answer, empty list if no provenances are present or answer is empty
-        dataset = dataset.map(lambda example: {'ranking_label': [[provenance['wikipedia_id'] for provenance in el['provenance']] if len(el['answer']) > 0 and len(el['provenance']) > 0 else [] for el in example['output']]})
-        dataset = dataset.rename_column("input", "content")
-        dataset = dataset.remove_columns(['meta', 'output'])
-        return dataset
-
-
-class KILTSTrex(Processor):
-
-    def __init__(self, *args, **kwargs):
-        dataset_name = 'kilt_trex'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    def process(self):
-        hf_name = 'kilt_tasks' 
-        hf_query_or_doc_name= "trex"
-        dataset = datasets.load_dataset(hf_name, hf_query_or_doc_name, num_proc=self.num_proc)[self.split]
-        # discarding empty answers 
-        dataset = dataset.map(lambda example: {'label': [el['answer'] for el in example['output'] if len(el['answer']) > 0]})
-        # ranking_label: list of wikipedia_ids per answer, empty list if no provenances are present or answer is empty
-        dataset = dataset.map(lambda example: {'ranking_label': [[provenance['wikipedia_id'] for provenance in el['provenance']] if len(el['answer']) > 0 and len(el['provenance']) > 0 else [] for el in example['output']]})
-        dataset = dataset.rename_column("input", "content")
-        dataset = dataset.remove_columns(['meta', 'output'])
-        return dataset
-    
-class KILTWned(Processor):
-
-    def __init__(self, *args, **kwargs):
-        dataset_name = 'kilt_wned'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    def process(self):
-        hf_name = 'kilt_tasks' 
-        hf_query_or_doc_name= "wned"
-        dataset = datasets.load_dataset(hf_name, hf_query_or_doc_name, num_proc=self.num_proc)[self.split]
-        # discarding empty answers 
-        dataset = dataset.map(lambda example: {'label': [el['answer'] for el in example['output'] if len(el['answer']) > 0]})
-        # ranking_label: list of wikipedia_ids per answer, empty list if no provenances are present or answer is empty
-        dataset = dataset.map(lambda example: {'ranking_label': [[provenance['wikipedia_id'] for provenance in el['provenance']] if len(el['answer']) > 0 and len(el['provenance']) > 0 else [] for el in example['output']]})
-        dataset = dataset.rename_column("input", "content")
-        dataset = dataset.remove_columns(['meta', 'output'])
-        return dataset
-
-class KILTWow(Processor):
-
-    def __init__(self, *args, **kwargs):
-        dataset_name = 'kilt_wow'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    def process(self):
-        hf_name = 'kilt_tasks' 
-        hf_query_or_doc_name= "wow"
-        dataset = datasets.load_dataset(hf_name, hf_query_or_doc_name, num_proc=self.num_proc)[self.split]
-        # discarding empty answers 
-        dataset = dataset.map(lambda example: {'label': [el['answer'] for el in example['output'] if len(el['answer']) > 0]})
-        # ranking_label: list of wikipedia_ids per answer, empty list if no provenances are present or answer is empty
-        dataset = dataset.map(lambda example: {'ranking_label': [[provenance['wikipedia_id'] for provenance in el['provenance']] if len(el['answer']) > 0 and len(el['provenance']) > 0 else [] for el in example['output']]})
-        dataset = dataset.rename_column("input", "content")
-        dataset = dataset.remove_columns(['meta', 'output'])
-        return dataset
     
 class MsMarcoQueries(Processor):
 
@@ -710,73 +214,6 @@ class MsMarcoQueries(Processor):
         queries_d = json.load(open("/gfs-ssd/user/tformal/neural_search/MSMARCO/dev_queries_collection/dev_queries.json"))  # super hard-coded path, see how to do properly
         ids, queries = zip(*queries_d.items())
         dataset = datasets.Dataset.from_dict({"id":ids, "content": queries})  # no need for split?
-        return dataset
-
-class MKQA(Processor):
-
-    def __init__(self, lang, *args, **kwargs):
-        dataset_name = f'mkqa_{lang}'
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-        self.lang = lang
-        
-    def process(self):
-        mkqa = datasets.load_dataset('mkqa')
-        kilt_nq = datasets.load_dataset("kilt_tasks", "nq")
-
-        mkqa_ids = {s['example_id']:i for i, s in enumerate(mkqa[self.split])}
-        kilt_nq_train_ids = {s['id']:i for i, s in enumerate(kilt_nq[self.split])}
-
-        overlap_ids = set(mkqa_ids.keys()).intersection(set(kilt_nq_train_ids.keys()))
-        overlap_mkqa = mkqa['train'].select([mkqa_ids[i] for i in overlap_ids])
-        overlap_kilt_nq = kilt_nq['train'].select([kilt_nq_train_ids[i] for i in overlap_ids])        
-        dataset = overlap_kilt_nq.add_column(f"content", [sample['queries'][self.lang] for sample in overlap_mkqa])    
-        # discarding empty answers
-        dataset = dataset.add_column(f"label", [[a['text'] for a in sample['answers'][self.lang] if not a['text']==None] for sample in overlap_mkqa])
-        # filter out samples with empty answer
-        dataset = dataset.filter(lambda example: len(example['label'])>0)
-
-        # ranking_label: list of wikipedia_ids per answer, empty list if no provenances are present or answer is empty
-        dataset = dataset.map(lambda example: {'ranking_label': [[provenance['wikipedia_id'] for provenance in el['provenance']] if len(el[f'answer']) > 0 and len(el['provenance']) > 0 else [] for el in example['output']]})        
-        dataset = dataset.remove_columns(['meta'])
-        return dataset
-
-class XORQA(Processor):
-
-    def __init__(self, lang, *args, **kwargs):
-        dataset_name = f'xor_tydiqa_{lang}'
-        self.lang = lang
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    def process(self):
-        def extend(label, lang):
-            if "yes" in label:
-                label += {"ru": ["да"], "ko": ["예"], "ja": ["はい"], "fi": ["kyllä", "joo"], "ar": ["نعم", "أجل", "بلى"]}[lang]
-            if "no" in label:
-                label += {"ru": ["нет"], "ko": ["아니요"], "ja": ["いいえ"], "fi": ["ei"], "ar": ["لا"]}[lang]
-            return label
-        os.system("wget https://nlp.cs.washington.edu/xorqa/XORQA_site/data/xor_dev_full_v1_1.jsonl")
-        dataset = datasets.load_dataset("json", data_files="xor_dev_full_v1_1.jsonl")["train"] # the file should be already .dev, and train is just default hf label
-        dataset = dataset.filter(lambda example: example['lang']==self.lang)
-        # discarding empty answers 
-        dataset = dataset.map(lambda example: {'label': extend([el for el in example['answers'] if len(el) > 0], self.lang)})
-        dataset = dataset.rename_column("question", "content")
-        #dataset = dataset.remove_columns(['meta', 'output'])
-        os.system("rm xor_dev_full_v1_1.jsonl")
-        return dataset
-
-class TydiQA(Processor):
-
-    def __init__(self, langcode="en", language="english", *args, **kwargs):
-        dataset_name = f'tydiqa_{langcode}'
-        self.language = language
-        super().__init__(*args, **kwargs, dataset_name=dataset_name)
-
-    def process(self):
-        dataset =  datasets.load_dataset("google-research-datasets/tydiqa", "secondary_task")[self.split] 
-        dataset = dataset.filter(lambda example: example['id'].startswith(self.language))
-        dataset = dataset.map(lambda example: {'label': [el for el in example['answers']["text"] if len(el) > 0]})
-        dataset = dataset.rename_column("question", "content")
-        dataset = dataset.remove_columns(['title', 'context', 'answers'])
         return dataset
 
 # ---------------------------------------- #
